requirements.txt
# 딥러닝 초해상도 화질 개선 및 모자이크 보정을 위한 Python 의존성

# PyTorch (CPU 버전 - GPU가 없어도 동작)
# GPU가 있는 경우, 아래 주석을 참고하여 CUDA 버전으로 설치하세요
torch>=2.0.0
torchvision>=0.15.0

# Real-ESRGAN (초해상도 모델)
realesrgan>=0.3.0

# 이미지 처리
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# 추가 이미지 처리 라이브러리
scikit-image>=0.21.0

# 참고: GPU 지원이 필요한 경우
# CUDA 11.8용 PyTorch 설치 예시:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# 
# CUDA 12.1용 PyTorch 설치 예시:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# 설치 후 다음 명령어로 확인:
# python -c "import torch; print(torch.cuda.is_available())"

